🛡️ Guardian (JULES): Consolidate DiffView components into src/components/ui/diff-view.tsx
WHAT:
- Merged `src/components/book/elements/diff-view.tsx` into `src/components/ui/diff-view.tsx`.
- Updated `DiffView` to support legacy MDX props (`before`, `after`) by mapping them to a new `split` layout mode.
- Migrated `VersionDiff` to the centralized core UI component.
- Deleted the redundant `src/components/book/elements/diff-view.tsx`.
- Updated barrel exports.
- Complied with JULES protocol for logging and pre-flight checks.

WHY:
- Eliminate redundant architectural smell: multiple string-diff components scattered across the codebase relying on identical internal LCS sequencing logic.
- Reduce code duplication and maintenance burden by maintaining a single source of truth for DiffView representations.

JULES CONTEXT:
- Checked `.Jules/task-log.md`: No Autonomous conflicts detected.
- Session logged: `.Jules/guardian/2026-06-18/session-report.md`.
- Reference: `.Jules/JULES.md`.

IMPACT:
- Files reduced: 2 -> 1.
- Code lines: Consolidated redundant `computeDiff` algorithms into one optimized `computeWordDiff` implementation.
- Test coverage: Maintained (all passed).

VERIFICATION:
- All tests pass: ✓
- TypeScript check: ✓
- Linting: ✓
- JULES compliance: ✓
