+++ .Jules/JULES.md
@@ -1,0 +1,4 @@
+# JULES (Joint Unified Logging and Execution Standard)
+1. All autonomous and persona tasks must log their start, progress, and end in .Jules/task-log.md.
+2. Before starting, read the task log to ensure no conflicts.
+3. Obey persona-specific boundaries.
