+++ .Jules/task-log.md
@@ -1,0 +1,10 @@
+# Task Log
+## 2026-06-18 16:33 - [GUARDIAN] Session started
+- Directory: .Jules/guardian/2026-06-18
+- Phase: PRE-FLIGHT
+- JULES Check: COMPLETE
+
+
+## 2026-06-18 16:47 - [GUARDIAN] Session completed
+- Consolidated: `src/components/book/elements/diff-view.tsx` into `src/components/ui/diff-view.tsx`
+- JULES Check: COMPLETE
