+++ src/components/book/elements/index.ts
@@ -33,7 +33,7 @@ export { TextToImageDemo, TextToVideoDemo } from "./media-demos";
 export { SummarizationDemo, ContextPlayground } from "./context-demos";
 export { BookPartsNav } from "./navigation";
 export { TokenPredictionDemo } from "./token-prediction";
-export { DiffView, VersionDiff } from "./diff-view";
+export { DiffView, VersionDiff } from "@/components/ui/diff-view";
 export { ChainErrorDemo } from "./chain-error-demo";
 export { ValidationDemo, FallbackDemo, ContentPipelineDemo } from "./chain-demos";
 export { FillInTheBlank, InteractiveChecklist, PromptDebugger } from "./exercises";
