+++ src/components/book/interactive.tsx
@@ -47,7 +47,7 @@ export { TextToImageDemo, TextToVideoDemo } from "./elements/media-demos";
 export { SummarizationDemo, ContextPlayground } from "./elements/context-demos";
 export { BookPartsNav } from "./elements/navigation";
 export { TokenPredictionDemo } from "./elements/token-prediction";
-export { DiffView, VersionDiff } from "./elements/diff-view";
+export { DiffView, VersionDiff } from "@/components/ui/diff-view";
 export { ChainErrorDemo } from "./elements/chain-error-demo";
 export { ValidationDemo, FallbackDemo, ContentPipelineDemo } from "./elements/chain-demos";
 export { FillInTheBlank, InteractiveChecklist, PromptDebugger } from "./elements/exercises";
